user delete done

+++ repositories/UserFavouriteRepository.js
@@ -1,69 +1,84 @@
-const { Op } = require('sequelize')
-const UserFavourite = require('../models/user_favourites')
-const Friends = require('../models/friends')
-const User = require('../models/user')
-const Chat = require('../models/chat')
+const { Op } = require("sequelize");
+const UserFavourite = require("../models/user_favourites");
+const Friends = require("../models/friends");
+const User = require("../models/user");
+const Chat = require("../models/chat");
 const sequelize = require("../config/database");
 
 class UserFavouriteRepository {
-  async create (userId, profileId) {
+  async create(userId, profileId) {
     const favourite = await UserFavourite.create({
       userId,
-      profileId
-    })
+      profileId,
+    });
     const chat = await Chat.findOne({
       where: {
         [Op.and]: [{ user2Id: profileId }, { user1Id: userId }],
       },
     });
-    return  { favourite, chat }
+    return { favourite, chat };
   }
 
-  async remove (userId, profileId) {
+  async remove(userId, profileId) {
     return UserFavourite.destroy({
       where: {
         userId,
-        profileId
-      }
-    })
-  }
-
-  async cancelUserFavourite(requesterId, requesteeId) {
-    return  await UserFavourite.destroy({
-      where: {
-        [Op.or]: [
-          { userId: requesterId },
-          { profileId: requesteeId },
-        ],
-      }
+        profileId,
+      },
     });
   }
 
-  async get (userId, profileId) {
+  async deleteUserFavourite(userId) {
+    const favourites = await UserFavourite.findAll({
+      where: {
+        [Op.or]: [{ userId: userId }, { profileId: userId }],
+      },
+    });
+
+    if (favourites.length > 0) {
+      await UserFavourite.destroy({
+        where: {
+          [Op.or]: [{ userId }, { profileId: userId }],
+        },
+      });
+    }
+
+    return favourites;
+  }
+
+  async cancelUserFavourite(requesterId, requesteeId) {
+    return await UserFavourite.destroy({
+      where: {
+        [Op.or]: [{ userId: requesterId }, { profileId: requesteeId }],
+      },
+    });
+  }
+
+  async get(userId, profileId) {
     const favourite = await UserFavourite.findOne({
       where: {
         userId,
-        profileId
-      }
-    })
+        profileId,
+      },
+    });
     return favourite ? favourite.toJSON() : null;
   }
 
-  async getFavourites (userId) {
+  async getFavourites(userId) {
     const userFavourites = await UserFavourite.findAll({
       where: {
-        userId
-      }
-    })
-    const favouriteIds = userFavourites.map(favourite => favourite.profileId)
+        userId,
+      },
+    });
+    const favouriteIds = userFavourites.map((favourite) => favourite.profileId);
     const users = await User.findAll({
       where: {
         id: {
-          [Op.in]: favouriteIds
-        }
+          [Op.in]: favouriteIds,
+        },
       },
       attributes: [
-        "id",             
+        "id",
         "firstName",
         "lastName",
         "username",
@@ -73,10 +88,10 @@ class UserFavouriteRepository {
         "profilePic",
         "description",
         "settings",
-        "phoneNumber"
+        "phoneNumber",
       ],
-      raw: true
-    })
+      raw: true,
+    });
     return users;
     // const friendship = await Friends.findAll({
     //   where: {
@@ -106,4 +121,4 @@ class UserFavouriteRepository {
   }
 }
 
-module.exports = UserFavouriteRepository
+module.exports = UserFavouriteRepository;
